Split up test cases
